Tidy scRNA2ChIP: move into its subdir, keep base dataset/datamodule clean
- Move BasescRNA2ChIPDataset into datasets/scRNA2ChIP/ (was at datasets root).
- Move scRNA2ChIPDataModule + _ChipCollator out of training/data_module.py
  (base module file) into datasets/scRNA2ChIP/scrna2chip_data_module.py next
  to the concrete Concat* classes. data_module.py base DataModule is untouched.
- base_rna_dataset: drop the **kwargs catch-all; BasescRNA2ChIPDataset now
  filters kwargs to the parent signature before super().__init__ so the upstream
  base dataset stays clean.
- training_config.__setstate__: drop the OT field pops; those features never
  produced checkpoints needing migration support.
