+++ bmfm_targets/datasets/base_rna_dataset.py
@@ -134,7 +134,6 @@ class BaseRNAExpressionDataset(Dataset):
         expose_zeros: str | None = None,
         limit_genes: list[str] | None = None,
         output_wrapper: Callable | None = None,
-        **kwargs,
     ) -> None:
         """
         Initializes the dataset.
